Closer to cw-multi-test
