+++ src/ibc/module/staking.rs
@@ -1,0 +1,5 @@
+use crate::StakeKeeper;
+
+use super::IbcModule;
+
+impl IbcModule for StakeKeeper {}
