+++ src/ibc/relayer/mod.rs
@@ -1,11 +1,15 @@
+//! Utils and traits to support IBC capabilites
+
 use anyhow::Result as AnyResult;
 use cosmwasm_std::{
-    Api, CustomMsg, CustomQuery, IbcPacketReceiveMsg, StdError, StdResult, Storage,
+    Api, Binary, CustomMsg, CustomQuery, IbcChannelCloseMsg, IbcChannelConnectMsg,
+    IbcChannelOpenMsg, IbcPacketAckMsg, IbcPacketReceiveMsg, IbcPacketTimeoutMsg, StdError,
+    StdResult, Storage,
 };
 use serde::de::DeserializeOwned;
 
 use crate::{
-    transactions::transactional, App, AppResponse, Bank, Distribution, Gov, Ibc, Module, Staking,
+    transactions::transactional, App, AppResponse, Bank, Distribution, Gov, Module, Staking,
     Stargate, Wasm,
 };
 
@@ -15,8 +19,14 @@ mod packet;
 pub use channel::{create_channel, create_connection, ChannelCreationResult};
 pub use packet::{relay_packet, relay_packets_in_tx, RelayPacketResult, RelayingResult};
 
-use super::{router::CosmosIbcRouter, IbcPacketRelayingMsg, IbcSimpleModule};
+use super::{
+    module::{IbcModule, IbcWasm},
+    types::MockIbcQuery,
+    IbcPacketRelayingMsg, IbcSimpleModule,
+};
 
+/// Gets the attribute value corresponding to an event
+/// Used to analyze IBC transactions
 pub fn get_event_attr_value(
     response: &AppResponse,
     event_type: &str,
@@ -37,6 +47,7 @@ pub fn get_event_attr_value(
     )))
 }
 
+/// Returns wether the event exists in the response
 pub fn has_event(response: &AppResponse, event_type: &str) -> bool {
     for event in &response.events {
         if event.ty == event_type {
@@ -46,6 +57,7 @@ pub fn has_event(response: &AppResponse, event_type: &str) -> bool {
     false
 }
 
+/// Gets all the attribute value for a specific event-attribute pair
 pub fn get_all_event_attr_value(
     response: &AppResponse,
     event: &str,
@@ -69,26 +81,67 @@ impl<BankT, ApiT, StorageT, CustomT, WasmT, StakingT, DistrT, GovT, StargateT>
 where
     CustomT::ExecT: CustomMsg + DeserializeOwned + 'static,
     CustomT::QueryT: CustomQuery + DeserializeOwned + 'static,
-    WasmT: Wasm<CustomT::ExecT, CustomT::QueryT>,
-    BankT: Bank,
+    WasmT: Wasm<CustomT::ExecT, CustomT::QueryT> + IbcWasm<CustomT::ExecT, CustomT::QueryT>,
+    BankT: Bank + IbcModule,
     ApiT: Api,
     StorageT: Storage,
     CustomT: Module,
-    StakingT: Staking,
+    StakingT: Staking + IbcModule,
     DistrT: Distribution,
     GovT: Gov,
     StargateT: Stargate,
 {
-    // Send any msg
-    pub(crate) fn relay(&mut self, msg: IbcPacketRelayingMsg) -> AnyResult<AppResponse> {
-        let block_info = &self.block_info();
+    /// Sends any relaying related message on the IBC module
+    pub fn relay(&mut self, msg: IbcPacketRelayingMsg) -> AnyResult<AppResponse> {
+        let App {
+            router,
+            api,
+            storage,
+            block,
+        } = self;
 
-        transactional(self.storage_mut(), |write_cache, _| {
+        transactional(storage, |write_cache, _| {
             // TODO, This also doesn't work because the app is borrowed mutably and immutably too many times
             // The only way it could work is with public cw-multi-test elements.
-            self.router
-                .ibc
-                .relay(&self.api, write_cache, &self.router, block_info, msg)
+            router.ibc.relay(&*api, write_cache, router, block, msg)
         })
     }
+
+    /// Queries the IBC module
+    pub fn ibc_query(&self, query: MockIbcQuery) -> AnyResult<Binary> {
+        let Self {
+            block,
+            router,
+            api,
+            storage,
+        } = self;
+
+        let querier = router.querier(api, storage, block);
+
+        router
+            .ibc
+            .general_query(api, storage, &querier, block, query)
+    }
+}
+
+/// This is added for modules to implement actions upon ibc actions.
+/// This kind of execution flow is copied from the WASM way of doing things and is not 100% completetely compatible with the IBC standard
+/// Those messages should only be called by the Ibc module.
+/// For additional Modules, the packet endpoints should be implemented
+/// The Channel endpoints are usually not implemented besides storing the channel ids
+#[cosmwasm_schema::cw_serde]
+pub enum IbcModuleMsg {
+    /// Open an IBC Channel (2 first steps)
+    ChannelOpen(IbcChannelOpenMsg),
+    /// Connect an IBC Channel (2 last steps)
+    ChannelConnect(IbcChannelConnectMsg),
+    /// Close an IBC Channel
+    ChannelClose(IbcChannelCloseMsg),
+
+    /// Receive an IBC Packet
+    PacketReceive(IbcPacketReceiveMsg),
+    /// Receive an IBC Acknowledgement for a packet
+    PacketAcknowledgement(IbcPacketAckMsg),
+    /// Receive an IBC Timeout for a packet
+    PacketTimeout(IbcPacketTimeoutMsg),
 }
